requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.11.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
psycopg2==2.9.1